requirements.txt
blinker                1.8.2
cachetools             5.3.3
click                  8.1.7
colorama               0.4.6
et-xmlfile             1.1.0
Flask                  3.0.3
Flask-WTF              1.2.1
itsdangerous           2.2.0
Jinja2                 3.1.4
logging                0.4.9.6
MarkupSafe             2.1.5
mysql-connector        2.2.9
mysql-connector-python 8.4.0
numpy                  2.0.0
openpyxl               3.1.4
pandas                 2.2.2
pip                    24.1.1
python-dateutil        2.9.0.post0
pytz                   2024.1
six                    1.16.0
tzdata                 2024.1
Werkzeug               3.0.3
WTForms                3.1.2